requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-types==0.6.0
    # via pydantic
anthropic==0.23.1
    # via agent-e (pyproject.toml)
anyio==4.3.0
    # via
    #   anthropic
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
autogen-agentchat==0.2.36
    # via agent-e (pyproject.toml)
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via
    #   pdfminer-six
    #   requests
click==8.1.7
    # via
    #   nltk
    #   typer
    #   uvicorn
cryptography==42.0.8
    # via pdfminer-six
diskcache==5.6.3
    # via
    #   autogen-agentchat
    #   pyautogen
distro==1.9.0
    # via
    #   anthropic
    #   openai
dnspython==2.6.1
    # via email-validator
docker==7.0.0
    # via
    #   autogen-agentchat
    #   pyautogen
email-validator==2.2.0
    # via fastapi
fastapi==0.111.1
    # via agent-e (pyproject.toml)
fastapi-cli==0.0.4
    # via fastapi
filelock==3.13.3
    # via huggingface-hub
flaml==2.1.1
    # via
    #   autogen-agentchat
    #   pyautogen
fsspec==2024.3.1
    # via huggingface-hub
google-ai-generativelanguage==0.6.2
    # via google-generativeai
google-api-core==2.19.0
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.127.0
    # via google-generativeai
google-auth==2.29.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-generativeai==0.5.1
    # via agent-e (pyproject.toml)
googleapis-common-protos==1.63.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.0.3
    # via playwright
grpcio==1.63.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via
    #   anthropic
    #   fastapi
    #   openai
    #   replicate
huggingface-hub==0.22.2
    # via tokenizers
idna==3.6
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.4
    # via fastapi
joblib==1.3.2
    # via nltk
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
nest-asyncio==1.6.0
    # via agent-e (pyproject.toml)
nltk==3.8.1
    # via agent-e (pyproject.toml)
numpy==1.26.4
    # via
    #   autogen-agentchat
    #   flaml
    #   pyautogen
openai==1.13.3
    # via
    #   autogen-agentchat
    #   pyautogen
packaging==23.2
    # via
    #   autogen-agentchat
    #   docker
    #   huggingface-hub
    #   replicate
pdfminer-six==20231228
    # via pdfplumber
pdfplumber==0.11.1
    # via agent-e (pyproject.toml)
pillow==10.3.0
    # via
    #   autogen-agentchat
    #   pdfplumber
playwright==1.44.0
    # via agent-e (pyproject.toml)
proto-plus==1.23.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.3
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pyautogen==0.2.27
    # via agent-e (pyproject.toml)
pycparser==2.22
    # via cffi
pydantic==2.6.2
    # via
    #   agent-e (pyproject.toml)
    #   anthropic
    #   autogen-agentchat
    #   fastapi
    #   google-generativeai
    #   openai
    #   pyautogen
    #   replicate
pydantic-core==2.16.3
    # via pydantic
pyee==11.1.0
    # via playwright
pygments==2.18.0
    # via rich
pyparsing==3.1.2
    # via httplib2
pypdfium2==4.30.0
    # via pdfplumber
python-dotenv==1.0.0
    # via
    #   agent-e (pyproject.toml)
    #   autogen-agentchat
    #   pyautogen
    #   uvicorn
python-json-logger==2.0.7
    # via agent-e (pyproject.toml)
python-multipart==0.0.9
    # via fastapi
pyyaml==6.0.1
    # via
    #   huggingface-hub
    #   uvicorn
regex==2023.12.25
    # via
    #   nltk
    #   tiktoken
replicate==0.34.1
    # via autogen-agentchat
requests==2.31.0
    # via
    #   docker
    #   google-api-core
    #   huggingface-hub
    #   tiktoken
rich==13.7.1
    # via typer
rsa==4.9
    # via google-auth
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
    #   httpx
    #   openai
starlette==0.37.2
    # via fastapi
tabulate==0.9.0
    # via agent-e (pyproject.toml)
termcolor==2.4.0
    # via
    #   autogen-agentchat
    #   pyautogen
tiktoken==0.6.0
    # via
    #   autogen-agentchat
    #   pyautogen
tokenizers==0.15.2
    # via anthropic
tqdm==4.66.2
    # via
    #   google-generativeai
    #   huggingface-hub
    #   nltk
    #   openai
typer==0.12.3
    # via fastapi-cli
typing-extensions==4.10.0
    # via
    #   anthropic
    #   fastapi
    #   google-generativeai
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   pyee
    #   replicate
    #   typer
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.2.1
    # via
    #   docker
    #   requests
uvicorn==0.30.3
    # via
    #   agent-e (pyproject.toml)
    #   fastapi
uvloop==0.19.0
    # via uvicorn
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
